Moved WebApi to .Net Core library; removed calling module project
