+++ EdgeLogging.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28010.2019
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EdgeLoggingServerModule", "EdgeLoggingServerModule\EdgeLoggingServerModule.csproj", "{E169A0B9-EBEB-49A0-B7FD-088F2D45D5A9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EdgeLoggingService", "EdgeLoggingService\EdgeLoggingService.csproj", "{447A4837-4D0C-4471-B346-8321114AF628}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EdgeLoggingService", "EdgeLoggingService\EdgeLoggingService.csproj", "{447A4837-4D0C-4471-B346-8321114AF628}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EdgeLoggingServer", "EdgeLoggingServer\EdgeLoggingServer.csproj", "{EC8DE060-C895-4DDC-818F-AFA880FE6C8E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E169A0B9-EBEB-49A0-B7FD-088F2D45D5A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E169A0B9-EBEB-49A0-B7FD-088F2D45D5A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E169A0B9-EBEB-49A0-B7FD-088F2D45D5A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E169A0B9-EBEB-49A0-B7FD-088F2D45D5A9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{447A4837-4D0C-4471-B346-8321114AF628}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{447A4837-4D0C-4471-B346-8321114AF628}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{447A4837-4D0C-4471-B346-8321114AF628}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{447A4837-4D0C-4471-B346-8321114AF628}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC8DE060-C895-4DDC-818F-AFA880FE6C8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC8DE060-C895-4DDC-818F-AFA880FE6C8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC8DE060-C895-4DDC-818F-AFA880FE6C8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC8DE060-C895-4DDC-818F-AFA880FE6C8E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
