--- EdgeLoggingServerModule/Dockerfile
@@ -1,30 +1,0 @@
-
-
-
-FROM microsoft/dotnet:2.1-sdk AS build-env
-WORKDIR /app
-
-COPY *.csproj ./
-RUN dotnet restore
-
-COPY . ./
-RUN dotnet publish -c Release -o out
-
-#FROM microsoft/dotnet:2.1-runtime
-FROM microsoft/dotnet:2.1.1-aspnetcore-runtime
-WORKDIR /app
-COPY --from=build-env /app/out ./
-
-
-
-
-ARG UID=1000
-RUN useradd -u $UID -ms /bin/bash moduleuser
-USER moduleuser
-
-# Expose ports
-EXPOSE 8888/tcp
-
-#VOLUME /app/config
-
-ENTRYPOINT ["dotnet", "FieldGatewayMicroservice.dll"]
