--- EdgeLoggingServerModule/EdgeLoggingServerModule.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="SkunkLab.Edge.Core" Version="1.0.47" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\EdgeLoggingService\EdgeLoggingService.csproj" />
-  </ItemGroup>
-
-</Project>
